refactor identity endpoints into modules
