+++ src/api/identity.rs
@@ -1,2 +1,4 @@
-pub mod requests;
-pub mod responses;
+pub mod entity;
+pub mod entity_alias;
+pub mod group;
+pub mod group_alias;
