+++ src/api/identity/entity.rs
@@ -1,0 +1,2 @@
+pub mod requests;
+pub mod responses;
